feat: 182015 Refactored project structure
