+++ openspp_dms_ftp_server/__init__.py
@@ -1,0 +1,5 @@
+from typing import TypeVar
+
+from pyftpdlib.handlers import FTPHandler
+
+TypeHandler = TypeVar("TypeHandler", bound=FTPHandler)
